GraphQL and indexer updates

+++ crates/myso-indexer-alt-social-schema/migrations/20260410200000_user_reservation_holdings_aggregate_amounts/up.sql
@@ -1,5 +1,6 @@
 -- user_reservation_holdings: spt_reservations.amount is a signed per-event delta (+deposit, -withdrawal),
 -- not a running balance. Aggregate SUM(amount) per (pool_id, reserver_address) for current holdings.
+-- SUM(bigint) is numeric; ::bigint keeps `amount` as bigint so CREATE OR REPLACE VIEW succeeds.
 CREATE OR REPLACE VIEW user_reservation_holdings AS
 SELECT
     agg.reserver_address,
@@ -17,7 +18,7 @@ FROM (
     SELECT
         reserver_address,
         pool_id,
-        SUM(amount) AS amount,
+        SUM(amount)::bigint AS amount,
         MAX(reserved_at) AS reserved_at
     FROM spt_reservations
     GROUP BY reserver_address, pool_id
